Functions Complement for LoRa Exceptions
This commit adds some functionality to manage exceptions in the LoRa feature. These exceptions work around issues that drop the transmission when the device is too far away or when the buffer is full.

The changed functions were OnRxError and OnCadDone in the Traveler file. The functions do a 100 ms delay and restart the channel activity detection, to try to send again, while in the Home file the functions are waiting for new data.