Code for some corrections and optimization
This code adds functions to LoRa_P2P to deal with exceptions like RX errors or buffer busy. Also, a description of each code is added in the README files.

README files have been added for each project. In LoRa_P2P/Home the OnCadDone and OnRxError functions were changed to handle exceptions. Finally, the way of printing the message was changed to make it clearer and more readable.